requirements.txt
attrs==25.3.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
folium==0.19.5
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-geo-type==0.3.12
google-maps-routing==0.6.15
googleapis-common-protos==1.69.2
googlemaps==4.10.0
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
numpy==2.2.4
polyline==2.0.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
typing_extensions==4.13.1
urllib3==2.3.0
xyzservices==2025.1.0
